refactor(modularization): Phase 4 Step 4-S4 (views) — marketplace routes merge to domain + app.py rewire (#25)
Views half of Phase 4 Step 4-S4 (services were moved in preceding
commit `85545f1d`).

**routes merge**:
- ``git mv`` ``aperag/views/marketplace.py`` →
  ``aperag/domains/marketplace/api/routes.py``; the former
  ``aperag/views/marketplace_collections.py`` content is merged into
  the same ``routes.py`` (two former router files collapse into one
  canonical domain router — same pattern Phase 3 Step 5a used for KB
  collections_v2 + documents_v2).
- Handlers keep identical URLs / methods / response_model / status
  codes — ``scripts/export_openapi.py --check`` clean (spec
  byte-stable).

**G1 boundary pass**:
- ``Depends(required_user)`` / ``Depends(optional_user)`` parameters
  typed as ``aperag.domains.marketplace.ports.AuthenticatedUser``
  instead of ``aperag.db.models.User`` (lesson 9a-ter + G16).
- ``aperag.schema.view_models.SharedCollection*`` imports switched to
  ``aperag.domains.marketplace.schemas``.
- ``aperag.schema.view_models.KnowledgeGraph`` switched to
  ``aperag.domains.knowledge_graph.schemas``.
- ``aperag.service.marketplace_service`` / ``marketplace_collection_service``
  imports switched to ``aperag.domains.marketplace.service.*``.
- ``aperag.service.graph_service`` switched to
  ``aperag.domains.knowledge_graph.service`` (already-moved Phase 2
  domain).
- ``aperag.service.document_service`` kept as **lazy import inside
  handler body** (same module reaches into KB domain only on the
  marketplace → KB subscriber read path; no top-level cross-domain
  binding).

**Q2 public rename call-site follow-up**:
- Four handler-body calls previously hitting
  ``marketplace_collection_service._check_marketplace_access(...)``
  now call the public ``check_marketplace_access(...)`` name per
  msg=6ab7d211.

**Legacy view shims**:
- ``aperag/views/marketplace.py`` and
  ``aperag/views/marketplace_collections.py`` collapse to
  router-import shims re-exporting the same merged ``APIRouter``
  instance (Phase 3 precedent). Two legacy imports resolve to the
  same router object.

**aperag/app.py rewire**:
- Direct import of the domain router
  (``from aperag.domains.marketplace.api.routes import router as
  marketplace_router``) replaces the two legacy
  ``aperag.views.marketplace{,_collections}`` imports.
- ``app.include_router(marketplace_router, prefix="/api/v1")`` is
  kept as a single include (not two) — the domain router carries
  both the marketplace-listing routes and the marketplace-collection
  read-only routes.

**FastAPI note**:
- The moved routes module intentionally does NOT use
  ``from __future__ import annotations`` — Phase 3 Step 5a + lesson
  9a-quatuordec canonical (status-204 handler + PEP 563 stringified
  ``-> Response`` annotation would crash
  ``is_body_allowed_for_status_code(204)``). This module has no 204
  routes today but the rule still applies for consistency.

**Verification**:
- ``scripts/export_openapi.py --check`` green — OpenAPI byte-stable.
- Full unit suite: 671 passed / 29 skipped.
- ``ruff check`` + ``ruff format --check`` clean.
- ``Grep "^from aperag\.(db\.models|service|schema\.view_models)"
  aperag/domains/marketplace/`` returns zero matches.

4-S4 now complete. 4-S5 (governance service + views move + Role
literal compare) is next.
