+++ aperag/app.py
@@ -47,6 +47,7 @@ from aperag.domains.knowledge_base.service.collection_service import (
     set_search_pipeline_ops as _kb_set_search_pipeline_ops,
 )
 from aperag.domains.knowledge_graph.api.routes import router as knowledge_graph_router
+from aperag.domains.marketplace.api.routes import router as marketplace_router
 from aperag.domains.retrieval.api.routes import router as retrieval_router
 from aperag.domains.web_access.api.routes import router as web_access_router
 from aperag.exception_handlers import register_exception_handlers
@@ -71,8 +72,6 @@ from aperag.views.evaluation_v2 import router as evaluation_v2_router
 from aperag.views.export import router as export_router
 from aperag.views.llm import router as llm_router
 from aperag.views.main import router as main_router
-from aperag.views.marketplace import router as marketplace_router
-from aperag.views.marketplace_collections import router as marketplace_collections_router
 from aperag.views.openai import router as openai_router
 from aperag.views.prompts import router as prompts_router
 from aperag.views.providers_v2 import router as providers_v2_router
@@ -130,8 +129,9 @@ app.include_router(export_router, prefix="/api/v1")  # Add export router
 app.include_router(api_key_router, prefix="/api/v1")
 app.include_router(audit_router, prefix="/api/v1")  # Add audit router
 app.include_router(llm_router, prefix="/api/v1")
-app.include_router(marketplace_router, prefix="/api/v1")  # Add marketplace router
-app.include_router(marketplace_collections_router, prefix="/api/v1")  # Add marketplace collections router
+app.include_router(
+    marketplace_router, prefix="/api/v1"
+)  # Marketplace domain router (marketplace + marketplace_collections)
 app.include_router(settings_router, prefix="/api/v1")
 app.include_router(prompts_router, prefix="/api/v1")  # Add prompts router
 app.include_router(web_access_router, prefix="/api/v2", tags=["web_access"])  # Add web_access domain router
